moved blog components into blog folder
