+++ src/templates/blogTemplate.js
@@ -1,6 +1,6 @@
 import React from "react"
 import { graphql } from "gatsby"
-import Post from "../components/Post"
+import Post from "../components/Blog/Post"
 
 export default function Template({
   data, // this prop will be injected by the GraphQL query below.
